requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncmy==0.2.9
babel==2.16.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.4
ghp-import==2.1.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
kafka==1.3.5
kafka-python-ng==2.2.3
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.46
mkdocs-material-extensions==1.3.1
packaging==24.2
paginate==0.5.7
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymdown-extensions==10.12
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==6.0.0
kafka-python
spark-sql-kafka
kafka-python-ng
mysql-connector-python

